Minor - Update 00_
We added some tips to deploy Git on a Windows machine (maybe the users that will need more help...).

We also removed several subsections regarding Git versioning usage (out of the scope of the bootcamp?):
 1) Remote Repositories
 2) Versioning
 3) Stash and Branch
 4) Git with Jupyter Notebooks

In next commit, we would like to add a small section presenting Stack Overflow (for sure most student will know it, but I think it is worth devoting a couple of statements).