Minor - Update 01_
We will start editing 01_ during upcoming days. I think I will drop some exercises for being a little bit too messy.

After talking with Lidia, we agree that maybe is better to change the tone of the notebook as a self-learning document for the students instead of master class style for the teachers.